Allow the second joint to be at 0 or 180 degrees
Arc cos can only be called on numbers from -1 to 1.
Before I was attempting to call it on numbers outside of this
range, causing an error.

This was particularly noticeable when trying to make
the arm was straight or with a 180 degree bend, since this wasn’t
possible- the argument passed into arc cos would have been very
slightly outside the range (e.g. 1.001) causing it to not draw.